+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Adding `xan count -c/--check-alignment`.
 * Adding `xan sort -e -z/--compress`.
 * Adding `xan sample -g -S/--sorted`.
+* Adding `xan from -f=(json|ndjson) --model`.
 
 *Fixes*